+++ src/nodeVisitor.hpp
@@ -62,23 +62,23 @@ Q_DECLARE_METATYPE(AnyNode)
 class NodeVisitor
 {
 public:
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::EntityNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::ConfigurationNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::AudioUnitNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::StreamInputNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::StreamOutputNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::AvbInterfaceNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::ClockSourceNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::LocaleNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::StringsNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::StreamPortNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::AudioClusterNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::AudioMapNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::ClockDomainNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::RedundantStreamNode const& node) noexcept = 0;
-	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::MemoryObjectNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::EntityNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::ConfigurationNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::AudioUnitNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::StreamInputNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::StreamOutputNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::AvbInterfaceNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::ClockSourceNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::LocaleNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::StringsNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::StreamPortNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::AudioClusterNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::AudioMapNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::ClockDomainNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::RedundantStreamNode const& node) noexcept = 0;
+	virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, la::avdecc::controller::model::MemoryObjectNode const& node) noexcept = 0;
 
-	static void accept(NodeVisitor* const visitor, la::avdecc::controller::ControlledEntity const* const entity, AnyNode const& node) noexcept;
+	static void accept(NodeVisitor* const visitor, la::avdecc::controller::ControlledEntity const* const entity, bool const isActiveConfiguration, AnyNode const& node) noexcept;
 
 	// Defaulted compiler auto-generated methods
 	NodeVisitor() noexcept = default;